harmonizes naming, refactors inputs menu into form components
